+++ edc-controlplane/edc-controlplane-base/build.gradle.kts
@@ -31,6 +31,7 @@ dependencies {
     runtimeOnly(project(":edc-extensions:edr:edr-callback"))
     runtimeOnly(project(":edc-extensions:tokenrefresh-handler"))
     runtimeOnly(project(":edc-extensions:agreements"))
+    runtimeOnly(project(":edc-extensions:validators:empty-asset-selector"))
 
     runtimeOnly(libs.edc.core.edrstore)
     runtimeOnly(libs.edc.edr.store.receiver)
@@ -92,7 +93,5 @@ dependencies {
     runtimeOnly(libs.edc.fc.core)
     runtimeOnly(libs.edc.fc.api)
 
-    //Has to be added last
-    runtimeOnly(project(":edc-extensions:validators:contractdefinitions"))
 
 }
