+++ edc-extensions/validators/empty-asset-selector/build.gradle.kts
