+++ edc-extensions/validators/empty-asset-selector/src/test/java/org/eclipse/tractusx/edc/validators/emptyassetselector/EmptyAssetSelectorValidatorTest.java
@@ -17,7 +17,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.edc.validators.contractdefinitions;
+package org.eclipse.tractusx.edc.validators.emptyassetselector;
 
 import jakarta.json.Json;
 import jakarta.json.JsonArrayBuilder;
