+++ settings.gradle.kts
@@ -53,7 +53,7 @@ include(":edc-extensions:dcp:tx-dcp")
 include(":edc-extensions:dcp:tx-dcp-sts-dim")
 include(":edc-extensions:data-flow-properties-provider")
 include(":edc-extensions:validators")
-include(":edc-extensions:validators:contractdefinitions")
+include(":edc-extensions:validators:empty-asset-selector")
 
 include(":edc-extensions:agreements")
 include(":edc-extensions:agreements:retirement-evaluation-core")
